refactor: PlaceTuripSelectionFragment 제거, TripDetailActivity 에서 처리하도록 수정
